+++ packages/elements-schema/package.json
@@ -1,0 +1,15 @@
+{
+    "name": "@sudoku-studio/elements-schema",
+    "version": "0.0.0",
+    "types": "./elementsSchema.d.ts",
+    "scripts": {
+        "clean": "rm -rf dist node_modules",
+        "check": "tsc --noEmit"
+    },
+    "devDependencies": {
+        "@sudoku-studio/schema": "workspace:*",
+        "@sudoku-studio/state-manager": "workspace:*",
+        "svelte": "^5.0.0",
+        "typescript": "^5.0.0"
+    }
+}
