+++ packages/elements-schema/tsconfig.json
@@ -1,0 +1,1 @@
+{ "extends": "../../tsconfig.base.json", "include": ["./*.ts"] }
