+++ packages/web/src/lib/js/element/selectDigitInputHandler.ts
@@ -1,7 +1,8 @@
 import { idxMapToKeysArray, cellCoord2CellIdx, cellIdx2cellCoord } from '@sudoku-studio/board-utils/src';
 import type { Geometry, Grid, Idx, IdxBitset, schema } from '@sudoku-studio/schema';
 import type { StateRef, Update } from '@sudoku-studio/state-manager/src';
-import { type GetInputHandler, inputHandler, adjacentCellPointerHandler } from '@sudoku-studio/elements/src';
+import type { GetInputHandler, InputHandler, InputHandlerContext } from '@sudoku-studio/elements-schema';
+import { adjacentCellPointerHandler } from '@sudoku-studio/elements-utils/src';
 import { boardState, getCellValue, getDigits } from '../board';
 import {
     MARK_TYPES,
@@ -13,6 +14,7 @@ import {
     userCursorIndexState,
     getUserToolStateName,
 } from '../user';
+import { parseDigit } from '@sudoku-studio/elements-utils/src';
 
 const selectPointerHandler = new adjacentCellPointerHandler.AdjacentCellPointerHandler(false);
 
@@ -36,15 +38,15 @@ export function makeSelectDigitGetInputHandler<V extends SelectDigitInputHandler
 }
 
 export function getSelectDigitInputHandler<V extends SelectDigitInputHandlerValue>(
-    ctx: inputHandler.InputHandlerContext<V>,
+    ctx: InputHandlerContext<V>,
     options: DigitInputHandlerOptions,
-): inputHandler.InputHandler {
+): InputHandler {
     const { multipleDigits, blockedByGivens, blockedByFilled, digitMapping, nextMode } = options;
 
     const DELETE_ORDER = ['filled', 'corner', 'center', 'colors'];
 
     function onDigitInput(code: string): boolean {
-        let digit: undefined | null | number | string = inputHandler.parseDigit(code);
+        let digit: undefined | null | number | string = parseDigit(code);
 
         if (digitMapping && null != digit && digit in digitMapping) {
             digit = digitMapping[digit];
